backend/requirements.txt
amqp==5.2.0
aniso8601==9.0.1
asgiref==3.7.2
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.0
Django==5.0.1
django-cors-headers==4.4.0
django-crontab==0.7.1
django-oauth-toolkit==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
graphene==3.3
graphene-django==3.2.2
graphql-core==3.2.4
graphql-relay==3.2.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.4
jwcrypto==1.5.6
kombu==5.4.0
MarkupSafe==2.1.5
numpy==2.0.0
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.1
pandas==2.2.2
pluggy==1.5.0
promise==2.3
prompt_toolkit==3.0.47
pycparser==2.22
PyJWT==2.9.0
pytest==8.3.2
pytest-django==4.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
setuptools==75.3.0
simplejson==3.19.3
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
